requirements.txt
requests
python-dotenv
polars
pandas
pyarrow